ui/requirements.txt
streamlit==1.44.0
requests==2.33.1
Pillow>=11.0.0
